requirements.txt
pytest
requests
jsonschema
allure-pytest
curlify